[milvus] Bump dependency 'etcd' to 8.2.3
This is the latest version of the chart that is available in the https repos.

The upgrade path from 6.3.3 to 8.2.3 has no noted compatability issues[1].
The major version change from 6->7 was moving back from 3.5.x to 3.4.x of
the underlying etcd app, and 7->8 was reinstating 3.5.x.

A number of new options have been added to the chart, mainly relating to
node selection of affinity settings.

[1] https://github.com/bitnami/charts/tree/main/bitnami/etcd#upgrading

Signed-off-by: Benedict Harcourt <ben.harcourt@harcourtprogramming.co.uk>